simpleem-poc/backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.59.5
python-dotenv==1.0.1
aiosqlite==0.20.0
aiofiles==24.1.0
pydantic==2.10.4
websockets==14.1
httpx==0.28.1
librosa>=0.10.0
numpy>=1.26.0
soundfile>=0.12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pyannote.audio>=3.1
torch>=2.0
fpdf2>=2.7
matplotlib>=3.8
